feat(issues): accept the command bar's Ctrl nav aliases in the thread panel
Ctrl+N/J move down and Ctrl+P/K move up, matching the command bar — cmdk
enables `vimBindings` by default and switches on exactly those four chords, so
the palette has always accepted them and the thread navigator was the odd one
out.

The policy already existed as pickerNavigationDirection, added for the editor's
mention and slash pickers for this same reason (MUL-5495). It lived in
suggestion-popup.tsx, which imports Tiptap and floating-ui, so reaching it from
a header popover would have dragged the editor's dependency graph along. Moved
both it and isPickerAcceptKey — pure KeyboardEvent predicates either way — to
common/picker-keys.ts, with their tests, and repointed the three consumers.

preventDefault matters for the letter aliases here specifically: focus sits in
the panel's search field, where Ctrl+K/N/P are readline editing commands that
would otherwise mangle the query while moving the cursor. Tab is deliberately
not an accept key in this panel, unlike the editor pickers, because Tab has to
stay focus navigation for the filter pills.

MUL-5755
